+++ src/main/java/com/notenoughmail/kubejs_tfc/KubeJSTFC.java
@@ -51,6 +51,7 @@ import java.util.stream.Stream;
 import static com.notenoughmail.kubejs_tfc.util.implementation.DataType.append;
 
 // TODO: 1.3.2 | Leaves, fallen leaves, tree solver command
+// TODO: [Future] | Custom recipe filters
 @SuppressWarnings("unused")
 @Mod(KubeJSTFC.MODID)
 public class KubeJSTFC {