+++ src/main/java/com/notenoughmail/kubejs_tfc/addons/firmalife/recipe/js/MixingBowlRecipeJS.java
@@ -1,11 +1,16 @@
 package com.notenoughmail.kubejs_tfc.addons.firmalife.recipe.js;
 
+import com.eerussianguy.firmalife.common.recipes.MixingBowlRecipe;
 import com.notenoughmail.kubejs_tfc.addons.firmalife.recipe.schema.MixingBowlSchema;
 import com.notenoughmail.kubejs_tfc.recipe.js.TFCRecipeJS;
 import dev.latvian.mods.kubejs.fluid.OutputFluid;
 import dev.latvian.mods.kubejs.item.InputItem;
 import dev.latvian.mods.kubejs.item.OutputItem;
+import dev.latvian.mods.kubejs.util.ConsoleJS;
 import net.dries007.tfc.common.recipes.ingredients.FluidStackIngredient;
+import net.minecraft.world.item.crafting.Ingredient;
+
+import java.util.List;
 
 @SuppressWarnings("unused")
 public class MixingBowlRecipeJS extends TFCRecipeJS {
@@ -40,5 +45,17 @@ public class MixingBowlRecipeJS extends TFCRecipeJS {
         setValue(MixingBowlSchema.INGREDIENTS, ingredients);
         setValue(MixingBowlSchema.FLUID_INGREDIENT, fluidIngredient);
         return this;
+    }
+
+    @Override
+    public List<Ingredient> getOriginalRecipeIngredients() {
+        if (getOriginalRecipe() instanceof MixingBowlRecipe m) {
+            return List.copyOf(m.getItemIngredients());
+        } else if (getOriginalRecipe() == null) {
+            ConsoleJS.SERVER.warn("Original recipe is null - could not get ingredients");
+            return List.of();
+        } else {
+            throw new IllegalStateException("Original recipe was not a mixing bowl recipe?");
+        }
     }
 }