+++ src/main/java/com/notenoughmail/kubejs_tfc/addons/firmalife/recipe/js/VatRecipeJS.java
@@ -1,5 +1,6 @@
 package com.notenoughmail.kubejs_tfc.addons.firmalife.recipe.js;
 
+import com.eerussianguy.firmalife.common.recipes.VatRecipe;
 import com.google.gson.JsonElement;
 import com.notenoughmail.kubejs_tfc.addons.firmalife.recipe.schema.VatSchema;
 import com.notenoughmail.kubejs_tfc.recipe.js.TFCProviderRecipeJS;
@@ -8,7 +9,12 @@ import com.notenoughmail.kubejs_tfc.util.implementation.ItemStackProviderJS;
 import dev.latvian.mods.kubejs.fluid.OutputFluid;
 import dev.latvian.mods.kubejs.item.InputItem;
 import dev.latvian.mods.kubejs.item.OutputItem;
+import dev.latvian.mods.kubejs.util.ConsoleJS;
 import net.dries007.tfc.common.recipes.ingredients.FluidStackIngredient;
+import net.dries007.tfc.common.recipes.ingredients.ItemStackIngredient;
+import net.minecraft.world.item.crafting.Ingredient;
+
+import java.util.List;
 
 @SuppressWarnings("unused")
 public class VatRecipeJS extends TFCProviderRecipeJS {
@@ -63,5 +69,17 @@ public class VatRecipeJS extends TFCProviderRecipeJS {
     public VatRecipeJS jar(OutputItem outputItem) {
         setValue(VatSchema.JAR, outputItem);
         return this;
+    }
+
+    @Override
+    public List<Ingredient> getOriginalRecipeIngredients() {
+        if (getOriginalRecipe() instanceof VatRecipe v) {
+            return v.getInputItem() == ItemStackIngredient.EMPTY ? List.of() : List.of(v.getInputItem().ingredient());
+        } else if (getOriginalRecipeResult() == null) {
+            ConsoleJS.SERVER.warn("Original recipe is null - could not get ingredients");
+            return List.of();
+        } else {
+            throw new IllegalStateException("Original recipe was not a vat recipe?");
+        }
     }
 }