+++ src/main/java/com/notenoughmail/kubejs_tfc/recipe/js/ChiselRecipeJS.java
@@ -4,6 +4,11 @@ import com.notenoughmail.kubejs_tfc.recipe.schema.ChiselSchema;
 import com.notenoughmail.kubejs_tfc.util.implementation.ItemStackProviderJS;
 import dev.latvian.mods.kubejs.item.InputItem;
 import dev.latvian.mods.kubejs.typings.Info;
+import dev.latvian.mods.kubejs.util.ConsoleJS;
+import net.dries007.tfc.common.recipes.ChiselRecipe;
+import net.minecraft.world.item.crafting.Ingredient;
+
+import java.util.List;
 
 // Technically has an ISP output
 @SuppressWarnings("unused")
@@ -19,5 +24,17 @@ public class ChiselRecipeJS extends TFCProviderRecipeJS {
     public ChiselRecipeJS extraDrop(ItemStackProviderJS extraDrop) {
         setValue(ChiselSchema.EXTRA_DROP, extraDrop);
         return this;
+    }
+
+    @Override
+    public List<Ingredient> getOriginalRecipeIngredients() {
+        if (getOriginalRecipe() instanceof ChiselRecipe c) {
+            return c.getItemIngredient() == null ? List.of() : List.of(c.getItemIngredient());
+        } else if (getOriginalRecipe() == null) {
+            ConsoleJS.SERVER.warn("Original recipe is null - could not get ingredients");
+            return List.of();
+        } else {
+            throw new IllegalStateException("Original recipe was not a chisel recipe?");
+        }
     }
 }