+++ src/main/java/com/notenoughmail/kubejs_tfc/recipe/js/SimplePotRecipeJS.java
@@ -1,12 +1,13 @@
 package com.notenoughmail.kubejs_tfc.recipe.js;
 
+import com.notenoughmail.kubejs_tfc.recipe.ISupportProviderOutput;
 import com.notenoughmail.kubejs_tfc.recipe.schema.SimplePotSchema;
 import com.notenoughmail.kubejs_tfc.util.implementation.ItemStackProviderJS;
 import dev.latvian.mods.kubejs.fluid.OutputFluid;
 import dev.latvian.mods.kubejs.typings.Info;
 
 @SuppressWarnings("unused")
-public class SimplePotRecipeJS extends TFCRecipeJS {
+public class SimplePotRecipeJS extends PotRecipeJS implements ISupportProviderOutput {
 
     @Info(value = "Sets the items that should be left in the pot after the recipe completes, accepts up to 5")
     public SimplePotRecipeJS itemOutput(ItemStackProviderJS[] itemOutput) {