+++ src/main/java/com/notenoughmail/kubejs_tfc/recipe/js/TFCRecipeJS.java
@@ -1,6 +1,36 @@
 package com.notenoughmail.kubejs_tfc.recipe.js;
 
 import dev.latvian.mods.kubejs.recipe.RecipeJS;
+import net.dries007.tfc.common.recipes.*;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.Recipe;
+
+import java.util.List;
 
 public class TFCRecipeJS extends RecipeJS {
+
+    @Override
+    public ItemStack getOriginalRecipeResult() {
+        return super.getOriginalRecipeResult();
+    }
+
+    @Override
+    public List<Ingredient> getOriginalRecipeIngredients() {
+        final Recipe<?> r = getOriginalRecipe();
+        if (r instanceof SimpleItemRecipe s) {
+            return List.of(s.getIngredient());
+        } else if (r instanceof BlastFurnaceRecipe b) {
+            return List.of(b.getCatalyst());
+        } else if (r instanceof CastingRecipe c) {
+            return List.of(c.getIngredient());
+        } else if (r instanceof IRecipeDelegate<?> d) {
+            return List.copyOf(d.getIngredients());
+        } else if (r instanceof GlassworkingRecipe g) {
+            return List.of(g.getBatchItem());
+        } else if (r instanceof WeldingRecipe w) {
+            return List.of(w.getFirstInput(), w.getSecondInput());
+        }
+        return super.getOriginalRecipeIngredients();
+    }
 }